[refactor] Extract `GridItemCellRenderControllerTests`
